增加食客端商家热门榜单页面（已完成）和商家详情页面（待完成）

+++ backend/src/main/java/com/foodadvisor/entity/ReviewVersion.java
@@ -18,10 +18,10 @@ public class ReviewVersion {
 
     private Long reviewId;
     private Integer version;
-    private Integer rating;
-    private Integer tasteRating;
-    private Integer environmentRating;
-    private Integer serviceRating;
+    private BigDecimal rating;
+    private BigDecimal tasteRating;
+    private BigDecimal environmentRating;
+    private BigDecimal serviceRating;
     private BigDecimal averageSpend;
     private LocalDate consumptionDate;
     private String content;
